fix: move popup code to device backend with other SDL funcs
